Fixed warnings

+++ chat.h
@@ -34,24 +34,24 @@ extern volatile sig_atomic_t running;
 void setup_signal_handler(void);
 
 // Pack a message with length prefix, returns total bytes to send
-static inline int pack_message(const char *msg, char *out_buf, int buf_size) {
-    int msg_len = strlen(msg);
+static inline ssize_t pack_message(const char *msg, char *out_buf, size_t buf_size) {
+    size_t msg_len = strlen(msg);
     if (msg_len + HEADER_SIZE > buf_size) return -1; // msg too big
 
-    uint32_t net_len = htonl(msg_len);
+    uint32_t net_len = htonl((unsigned int)msg_len);
     memcpy(out_buf, &net_len, HEADER_SIZE);
     memcpy(out_buf + HEADER_SIZE, msg, msg_len);
 
-    return msg_len + HEADER_SIZE;
+    return (ssize_t)(msg_len + HEADER_SIZE);
 }
 
 // Unpack a length-prefixed message. Returns message length or -1 on error
-static inline int unpack_message(int fd, char *out_msg, int max_len) {
+static inline ssize_t unpack_message(int fd, char *out_msg, size_t max_len) {
     uint32_t net_len;
-    int msg_len;
+    size_t msg_len;
 
     // Read 4-byte length header
-    int n = recv(fd, &net_len, HEADER_SIZE, 0);
+    ssize_t n = recv(fd, &net_len, HEADER_SIZE, 0);
     if (n <= 0) return n;  // 0 = disconnect, -1 = error
     if (n < HEADER_SIZE) return -1;  // Incomplete header
 
@@ -59,26 +59,26 @@ static inline int unpack_message(int fd, char *out_msg, int max_len) {
     if (msg_len <= 0 || msg_len >= max_len) return -1;
 
     // Read the actual message
-    int total = 0;
-    while (total < msg_len) {
-        n = recv(fd, out_msg + total, msg_len - total, 0);
+    ssize_t total = 0;
+    while (total < (ssize_t)msg_len) {
+        n = recv(fd, out_msg + total, msg_len - (size_t)total, 0);
         if (n <= 0) return n;
         total += n;
     }
     out_msg[msg_len] = '\0';
 
-    return msg_len;
+    return (ssize_t)msg_len;
 }
 
 // Send a length-prefixed message
-static inline int send_message(int fd, const char *msg) {
+static inline ssize_t send_message(int fd, const char *msg) {
     char packet[MAX_MSG_LEN + HEADER_SIZE];
-    int total = pack_message(msg, packet, sizeof(packet));
+    ssize_t total = pack_message(msg, packet, sizeof(packet));
     if (total < 0) return -1;
 
-    int sent = 0;
+    ssize_t sent = 0;
     while (sent < total) {
-        int n = send(fd, packet + sent, total - sent, 0);
+        ssize_t n = send(fd, packet + sent, (size_t)(total - sent), 0);
         if (n <= 0) return n;
         sent += n;
     }
